g!mt comment message

+++ commands/trading/multiTrade.js
@@ -8,7 +8,7 @@ function makeEmbed(user, otherUser, userTrade, otherUserTrade, checkUser) {
     const tradeEmebed = new EmbedBuilder()
         .setColor("#616161")
         .setTitle(`Multiple Trade [${user.username} <--> ${otherUser.username}]`)
-        .setDescription(`**${user}\'s items:**${checkUser == 0 ? " ✅" : ""}\n\`\`\`diff\n${userTrade.length ? userTrade.join("\n") : "- No items seleceted..."}\`\`\`\n**${otherUser}\'s items:**${checkUser == 1  ? " ✅" : ""}\n\`\`\`diff\n${otherUserTrade.length ? otherUserTrade.join("\n") : "- No items seleceted..."}\`\`\`\nTo add a \`Card\`, type its card code. **EX:** aa0000\nTo remove a \`Card\`, type its card code again. **EX:** aa0000\n\nTo add an \`Item\`, type its name followed it its amount.**EX:** Draw 5 x2\nTo remove an \`Item\`, type its name and its amount as 0. **EX:** Draw 5 x0\n*The \"x\" in front of the amount is required*\n\nTo add multiple cards or items at a time, sperate each thing using commas. \n**EX:** aa0000, aa0001, Draw 5 x2`)
+        .setDescription(`**${user}\'s items:**${checkUser == 0 ? " ✅" : ""}\n\`\`\`diff\n${userTrade.length ? userTrade.join("\n") : "- No items seleceted..."}\`\`\`\n**${otherUser}\'s items:**${checkUser == 1  ? " ✅" : ""}\n\`\`\`diff\n${otherUserTrade.length ? otherUserTrade.join("\n") : "- No items seleceted..."}\`\`\`\nTo add a \`Card\`, type its card code. **EX:** aa0000\nTo remove a \`Card\`, type its card code again. **EX:** aa0000\n\nTo add an \`Item\`, type its name followed it its amount.**EX:** Draw 5 x2\nTo remove an \`Item\`, type its name and its amount as 0. **EX:** Draw 5 x0\n*The \"x\" in front of the amount is required*\n\nTo add multiple cards or items at a time, sperate each thing using commas. \n**EX:** aa0000, aa0001, Draw 5 x2\n\nTo say something with out the bot deleting your message, start your message with a "#". **EX:** #Accept the trade by pressing the blue check mark button.`)
     return tradeEmebed;
 }
 
@@ -208,6 +208,8 @@ module.exports = {
             else { return; }
 
             const splitM = m.content.split(",");
+            if (splitM[0][0] == "#") { return; }
+
             otherConfirmUser = null;
             let checkUser = -1;
             m.delete();
